backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
redis==6.4.0
cryptography==45.0.7
pillow==11.3.0
pytesseract==0.3.13
numpy==2.3.3
torch==2.8.0
torchvision==0.23.0
email-validator==2.3.0
